requirements.txt
discord.py
asyncio
typing
Pillow
dismusic
PyNaCl
pyjokes